Fixed CDS slicing bug.
Converting from a list of ranges to a list of boundaries and using
`bisect_left()` was broken, this is only noticeable for very small exons. A
binary search algorithm that, given a coordinate, finds the nearest location
was implemented to resolve this issue. A related minor issue, assigninment of
the center of an intron in an inverted `MultiLocus` was broken. The new search
algorithm has a parameter `p` to deal with this issue.

Using `_cut()` on a boundary was broken, it resulted in generating a location
of length zero. The `_cut()` function was modified to filter out these
problematic locations.

A `Crossmap` instance containing a CDS, but without a 5'- or 3' UTR was broken.
An empty UTR can now detected using the modified `_cut()` function, in this
case the `MultiLocus` instance for the CDS is used.